bkpr: move json_to functions to respective type files
