+++ plugins/bkpr/chain_event.h
@@ -2,11 +2,13 @@
 #define LIGHTNING_PLUGINS_BKPR_CHAIN_EVENT_H
 
 #include "config.h"
+#include <bitcoin/tx.h>
 #include <ccan/short_types/short_types.h>
 
 struct amount_msat;
 struct bitcoin_outpoint;
 struct bitcoin_txid;
+struct json_stream;
 
 struct chain_event {
 
@@ -52,5 +54,8 @@ struct chain_event {
 	/* Sometimes chain events resolve payments */
 	struct sha256 *payment_id;
 };
+
+void json_add_chain_event(struct json_stream *out,
+                          struct chain_event *ev);
 
 #endif /* LIGHTNING_PLUGINS_BKPR_CHAIN_EVENT_H */
